split schema and table from summarize
